Adapted staticfile generation and added kshttp
Changed generation mechanism for static files in kshttp to have strings
that are discontinued at the ending of a line to have shorter string
constants that are concatenated by the compiler.
Build mechanism for static files was integrated into CMake Build.
Working directory for static files is now an argument to the script.